img-vit/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==20.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleu==0.3
blis==0.7.11
Bottleneck==1.2.1
catalogue==2.0.10
cephfs==2.0.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==8.1.7
cloudpathlib==0.16.0
ClusterShell==1.8.3
colorama==0.4.6
comm==0.2.2
confection==0.1.4
cryptography==3.3.2
cycler==0.10.0
cymem==2.0.8
Cython==0.29.21
datasets==2.19.0
dbus-python==1.2.16
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.8
distlib==0.3.1
distro-info==1.0+deb11u1
docker-pycreds==0.4.0
efficiency==2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.0.1
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.0.12
frozenlist==1.4.1
fsspec==2024.3.1
GDAL==3.2.2
gensim==3.8.3
git-cola==3.9
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
html5lib==1.1
httplib2==0.18.1
huggingface-hub==0.23.2
idna==2.10
imageio==2.34.1
importlib-metadata==1.6.0
iniconfig==0.0.0
invoke==1.4.1
iotop==0.6
ipykernel==6.29.4
ipython==7.20.0
ipython_genutils==0.2.0
ipywidgets==8.1.2
isc==2.0
jdcal==1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.3.1
langcodes==3.3.0
lazy_loader==0.4
libvirt-python==7.0.0
lightning-utilities==0.11.2
llvmlite==0.35.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.6
mock==4.0.3
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mysqlclient==1.4.4
nest-asyncio==1.6.0
networkx==2.5
nlg-eval @ git+https://github.com/Maluuba/nlg-eval.git@2ab4528fad5548315cf61e40c2249fec8c8ad233
nlp==0.4.0
nltk==3.8.1
numba==0.52.0
numexpr==2.7.2
numpy==1.19.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
odfpy==1.4.1
olefile==0.46
opencv-python==4.9.0.80
openpyxl==3.0.3
packaging==20.9
pandas==1.1.5
paramiko==2.7.2
parso==0.8.1
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
platformdirs==4.2.0
pluggy==0.13.0
ply==3.11
portalocker==2.8.2
preshed==3.0.9
prettytable==0.7.2
prompt-toolkit==3.0.14
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.8.6
pure-eval==0.2.2
py==1.10.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyclean==3.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.20
pydantic==2.6.4
pydantic_core==2.16.3
pydot==1.4.2
Pygments==2.7.1
PyGObject==3.38.0
pygraphviz==1.7
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytest==6.0.2
python-apt==2.2.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
pyzmq==25.1.2
QtPy==1.9.0
rados==2.0.0
rbd==2.0.0
regex==2023.12.25
requests==2.25.1
responses==0.18.0
rouge==1.0.1
sacrebleu==2.4.2
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.6.0
sentry-sdk==2.1.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.2.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tables==3.6.1
tabulate==0.9.0
Theano==1.0.5
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.4.24
tokenizers==0.19.1
toml==0.10.1
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.0.5
transformers==4.41.2
triton==2.2.0
typer==0.9.4
typing_extensions==4.11.0
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.4.0+ds
wandb==0.16.6
wasabi==1.1.2
wcwidth==0.1.9
weasel==0.3.4
webencodings==0.5.1
widgetsnbextension==4.0.10
xdg==6.0.0
xlwt==1.3.0
xxhash==3.4.1
yarl==1.9.4
zipp==1.0.0
